Moved code to legacy package
