--- src/main/java/org/skaggsm/minesweeperj/entities/Player.java
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2017 Mitchell Skaggs
- *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU Lesser General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     GNU Lesser General Public License for more details.
- *
- *     You should have received a copy of the GNU Lesser General Public License
- *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.skaggsm.minesweeperj.entities;
-
-import org.skaggsm.minesweeperj.board.Move;
-import org.skaggsm.minesweeperj.board.View;
-
-/**
- * This class defines the identity of player of a game.
- *
- * @author Mitchell Skaggs
- */
-public interface Player<M extends Move, V extends View> extends Viewer<V>, Participant<M, V> {
-    String getName();
-}
